Suggestions from @datrio
If you'll like House of Cards, you should also like The West Wing. Spartacus is great if you're into more action-like series. The Walking Dead might be boring at times (personal opinion), but most often you'll be eagerly waiting for the next ep. And Cowboy Bebop is an anime, but it's critically acclaimed and I don't know of a person who didn't like it, so I still think you should consider it.